add screen shots to q2 and q3

+++ q2_t1/q2_t11.c
@@ -2,7 +2,7 @@
 #include <math.h>
 #include <stdlib.h>
 
-long double poisson(int k ,double lambda) {
+long double poisson(int k ,int lambda) {
     return (100*(exp(-lambda) * pow(lambda, k)) / tgamma(k + 1));
 }
